+++ nls/de/strings.js
@@ -2,14 +2,15 @@ define({
     BEAUTIFY_ON_SAVE: 'Beautify beim Speichern',
     BEAUTIFY_ON_SAVE_DESC: 'Aktiviert Beautify beim Speichern',
 
+    PREF_BEAUTIFIERS_NAME: 'Externe Beautifier',
+    PREF_BEAUTIFIERS_DESC: 'Externe Beautifier, siehe https://github.com/brackets-beautify/brackets-beautify#external-formatters',
+    PREF_LANGUAGES_NAME: 'Beautifier-Sprachen-Zuordnung',
+    PREF_LANGUAGES_DESC: 'Beautifier-Sprachen-Zuordnung, siehe https://github.com/brackets-beautify/brackets-beautify#configure-languages',
+
     PREF_DIALOG_NAME: 'Zeige keinen Dialog für nicht unterstüzte Sprachen',
     PREF_DIALOG_DESC: 'Aktivieren, um keinen Dialog für nicht unterstüzte Sprachen zu zeigen',
     UNSUPPORTED_TITLE: 'Sprache nicht unterstützt',
     UNSUPPORTED_MESSAGE: 'Diese Sprache ist nicht unterstützt.<br />Verfügbare Sprachen sind JavaScript, JSON, HTML, XML, SVG, HTML in PHP, Embedded JavaScript, Handlebars, CSS, SCSS, und LESS.',
-    UNSUPPORTED_CHECKBOX: 'Zeige diese Meldung zukünftig nicht mehr',
-
-    SASS_PREF_NAME: 'Pfad zum sass-convert Programm',
-    SASS_PREF_DESC: 'Absoluter Pfad zum sass-convert Programm',
-    SASS_FORMAT: 'Fehler beim formatieren der SASS Datei',
-    SASS_ERROR: 'Bitte gib einen Pfad zum SASS-Convert Programm an.'
+    CONFIGURE_LINKS: 'In der README stehen informationen, um <a href="https://github.com/brackets-beautify/brackets-beautify#configure-languages">Sprachen zu konfigurieren</a> und <a href="https://github.com/brackets-beautify/brackets-beautify#external-formatters">neue Formattierer hinzuzufügen</a>.',
+    UNSUPPORTED_CHECKBOX: 'Zeige diese Meldung zukünftig nicht mehr'
 });