+++ nls/es/strings.js
@@ -1,9 +1,6 @@
 define({
     BEAUTIFY: 'Formatear documento',
-    BEAUTIFY_ON_SAVE: 'Formato automatico al guardar',
-
-    SASS_FORMAT: 'Ha ocurrido un error al formatear el archivo SASS',
-    SASS_ERROR: 'Debe especificar la ruta al programa de conversión SASS'
+    BEAUTIFY_ON_SAVE: 'Formato automatico al guardar'
 });
 
 /* Last translated for <VERSION> (<COMMIT-HASH>) */