+++ nls/it/strings.js
@@ -1,10 +1,7 @@
 define({
     BEAUTIFY_ON_SAVE: 'Beautify al salvataggio',
     UNSUPPORTED_TITLE: 'Linguaggio non supportato',
-    UNSUPPORTED_MESSAGE: 'Questo linguaggio non é supportato.<br />I linguaggi supportati sono: JavaScript, JSON, HTML, XML, SVG, HTML in PHP, Embedded JavaScript, Handlebars, CSS, SCSS e LESS.',
-
-    SASS_FORMAT: 'Si è verificato un errore di formattazione del file SASS',
-    SASS_ERROR: 'È necessario fornire un percorso del programma convertitore SASS'
+    UNSUPPORTED_MESSAGE: 'Questo linguaggio non é supportato.<br />I linguaggi supportati sono: JavaScript, JSON, HTML, XML, SVG, HTML in PHP, Embedded JavaScript, Handlebars, CSS, SCSS e LESS.'
 });
 
 /* Last translated for v.1.0.2 (a4466ce026f101107a3fab4fc371194a47868473) */