+++ nls/ro/strings.js
@@ -7,11 +7,6 @@ define({
     PREF_DIALOG_DESC: 'true pentru a ascunde avertizările legate de limbajele nesuportate',
     UNSUPPORTED_TITLE: 'Limbaj nesuportat',
     UNSUPPORTED_MESSAGE: 'Acest limbaj nu este suportat.<br />Limbajele suportate sunt JavaScript, JSON, HTML, XML, SVG, HTML în PHP, Embedded JavaScript, Handlebars, CSS, SCSS și LESS.',
-    UNSUPPORTED_CHECKBOX: 'Ascunde acest dialog pe viitor',
-
-    SASS_PREF_NAME: 'Calea către sass-convert',
-    SASS_PREF_DESC: 'Calea absolută către sass-convert',
-    SASS_FORMAT: 'Eroare in timpul formatării fișierului SASS',
-    SASS_ERROR: 'Trebuie să specifici calea către sass-convert'
+    UNSUPPORTED_CHECKBOX: 'Ascunde acest dialog pe viitor'
 });