+++ nls/root/strings.js
@@ -3,14 +3,15 @@ define({
     BEAUTIFY_ON_SAVE: 'Beautify on Save',
     BEAUTIFY_ON_SAVE_DESC: 'true to Beautify file on Save',
 
+    PREF_BEAUTIFIERS_NAME: 'External Beautifiers',
+    PREF_BEAUTIFIERS_DESC: 'External Beautifiers, see https://github.com/brackets-beautify/brackets-beautify#external-formatters',
+    PREF_LANGUAGES_NAME: 'Beautifier-Language-Mapping',
+    PREF_LANGUAGES_DESC: 'Beautifier-Language-Mapping, see https://github.com/brackets-beautify/brackets-beautify#configure-languages',
+
     PREF_DIALOG_NAME: 'Hide dialog for unsupported languages',
     PREF_DIALOG_DESC: 'true to hide the dialog warning about unsupported languages',
     UNSUPPORTED_TITLE: 'Unsupported Language',
     UNSUPPORTED_MESSAGE: 'This language is not supported.<br />Supported languages are JavaScript, JSON, HTML, XML, SVG, HTML in PHP, Embedded JavaScript, Handlebars, CSS, SCSS, and LESS.',
-    UNSUPPORTED_CHECKBOX: 'Don\'t show this dialog in the future',
-
-    SASS_PREF_NAME: 'Path to sass-convert',
-    SASS_PREF_DESC: 'Absolute path to sass-convert',
-    SASS_FORMAT: 'An error occurred formatting the SASS file',
-    SASS_ERROR: 'You need to provide a path to the sass-convert program'
+    CONFIGURE_LINKS: 'See the README for <a href="https://github.com/brackets-beautify/brackets-beautify#configure-languages">configuring languages</a> and <a href="https://github.com/brackets-beautify/brackets-beautify#external-formatters">adding new formatters</a>.',
+    UNSUPPORTED_CHECKBOX: 'Don\'t show this dialog in the future'
 });