+++ nls/zh-cn/strings.js
@@ -7,10 +7,5 @@ define({
     PREF_DIALOG_DESC: '勾选以隐藏不支持的语言警告对话框',
     UNSUPPORTED_TITLE: '不支持的语言',
     UNSUPPORTED_MESSAGE: '不支持此语言。<br />支持的语言如下： JavaScript, JSON, HTML, XML, SVG, HTML in PHP, 内嵌JavaScript, Handlebars, CSS, SCSS, LESS。',
-    UNSUPPORTED_CHECKBOX: '不再显示此对话框',
-
-    SASS_PREF_NAME: 'sass-convert的路径',
-    SASS_PREF_DESC: 'sass-convert的绝对路径',
-    SASS_FORMAT: '格式化SASS文件时出错',
-    SASS_ERROR: '需要提供sass-convert程序路径'
+    UNSUPPORTED_CHECKBOX: '不再显示此对话框'
 });